pyproject.toml
[project]
name = "m-anim-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.18.0",
    "manim>=0.19.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
